:truck: 이종수: 기존 데이터 조작어 기능 관련한 파일들, 폴더 세분화
