--- src/DB/modules/manipulation/CreateData.interface.ts
@@ -1,3 +1,0 @@
-import ICreateRecord from "../../interfaces/manipulation/CreateRecord.interface";
-
-export default interface ICreateData extends ICreateRecord {}
