--- src/DB/modules/manipulation/DeleteData.interface.ts
@@ -1,4 +1,0 @@
-import IDeleteRecord from "../../interfaces/manipulation/DeleteRecord.interface";
-import IDeleteRecordsAll from "../../interfaces/manipulation/DeleteRecordsAll.interface";
-
-export default interface IDeleteData extends IDeleteRecord, IDeleteRecordsAll {}
