--- src/DB/modules/manipulation/ReadData.interface.ts
@@ -1,8 +1,0 @@
-import IReadRecord from "../../interfaces/manipulation/ReadRecord.interface";
-import IReadRecordsAll from "../../interfaces/manipulation/ReadRecordsAll.interface";
-import IReadRecordsAllByIndex from "../../interfaces/manipulation/ReadRecordsAllByIndex.interface";
-
-export default interface IReadData
-  extends IReadRecord,
-    IReadRecordsAll,
-    IReadRecordsAllByIndex {}
