--- src/DB/modules/manipulation/UpdateData.interface.ts
@@ -1,3 +1,0 @@
-import IUpdateRecord from "../../interfaces/manipulation/UpdateRecord.interface";
-
-export default interface IUpdateData extends IUpdateRecord {}
