+++ src/DB/modules/manipulation/create/CreateData.interface.ts
@@ -1,0 +1,3 @@
+import ICreateRecord from "../../../interfaces/manipulation/CreateRecord.interface";
+
+export default interface ICreateData extends ICreateRecord {}
