+++ src/DB/modules/manipulation/delete/DeleteData.interface.ts
@@ -1,0 +1,4 @@
+import IDeleteRecord from "../../../interfaces/manipulation/DeleteRecord.interface";
+import IDeleteRecordsAll from "../../../interfaces/manipulation/DeleteRecordsAll.interface";
+
+export default interface IDeleteData extends IDeleteRecord, IDeleteRecordsAll {}
