+++ src/DB/modules/manipulation/read/ReadData.ts
@@ -1,5 +1,5 @@
-import DBConnector from "../../DBMANAGER";
-import instanceChecker from "../../throw/instanceChecker";
+import DBConnector from "../../../DBMANAGER";
+import instanceChecker from "../../../throw/instanceChecker";
 import IReadData from "./ReadData.interface";
 
 abstract class AbstractReadData extends DBConnector implements IReadData {
