+++ src/DB/modules/manipulation/update/UpdateData.interface.ts
@@ -1,0 +1,3 @@
+import IUpdateRecord from "../../../interfaces/manipulation/UpdateRecord.interface";
+
+export default interface IUpdateData extends IUpdateRecord {}
